Remove Client.portOff
There is no known reason for the Client to have a port offset. A port offset is used so that multiple services may be running on the same box (allowing each to be addressed individually). Since the Client is not the service and is not connected to by other Clients it does not need a port offset.

Also, the port offset was never assigned a value other than 0 which effectively provided no offset from the original port number.